requirements.txt
beautifulsoup4
html5lib
requests
python-telegram-bot
python-dotenv
pydash
lxml
python-dateutil